Created ProductServiceTests with Mock

+++ backend/src/main/java/com/devsuperior/dscatalog/services/ProductService.java
@@ -63,6 +63,10 @@ public class ProductService {
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Resource not found");
+        }
+
         try {
             repository.deleteById(id);
         } catch (DataIntegrityViolationException e) {
